+++ CHANGELOG.md
@@ -167,6 +167,19 @@
   the cluster's IP family from the `default/kubernetes` Service at startup. If
   detection failed, please set `--ip-family` manually.
   [#5499](https://github.com/Kong/kong-operator/pull/5499)
+- Hybrid Gateway: Support `konghq.com/konnect-cp-labels` and `konghq.com/konnect-dp-labels`
+  annotations in `Gateway`s and `GatewayClass`es to configure Konnect labels of
+  control planes and dataplanes. The key and value in the label are separated by
+  `=` and multiple labels in an annotation are separated by `,`.
+  For example: `konghq.com/konnect-cp-labels: "k1=v1,k2=v2"`.
+  The maximum number of labels is limited to 5. If the number of labels in
+  `konghq.com/konnect-cp-labels` or `konghq.com/konnect-dp-labels` annotation is
+  over 5 in `Gateway` or `GatewayClass`, or the number of combined labels from
+  the two parts is over 5, the controller rejects the `Gateway`.
+  The labels from annotations of `Gateway`s take precedence over the labels from
+  `GatewayClass`es. When they have different values on the same key, the value
+  from annotation of `Gateway` is used.
+  [#5808](https://github.com/Kong/kong-operator/pull/5808)
 
 ### Breaking changes